requirements.txt

streamlit==1.42.0
tensorflow==2.19.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==2.0.2
supabase==2.24.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.3
gdown==4.7.3
python-dotenv==1.0.0
fpdf2==2.7.9
